chore: Reorganize guild related SWR hooks.
