+++ src/components/Guilds/Filter/index.tsx
@@ -10,9 +10,9 @@ import { Box, Flex } from '../common/Layout/Box';
 import { FilterMenu, FilterButton, FilterBadge } from './FilterMenu';
 import { Button } from '../common/Button';
 import { useHistory, useLocation } from 'react-router';
-import { useVotingPowerOf } from 'hooks/Guilds/ether-swr/useVotingPowerOf';
+import { useVotingPowerOf } from 'hooks/Guilds/ether-swr/guild/useVotingPowerOf';
 import { useWeb3React } from '@web3-react/core';
-import { useGuildConfig } from 'hooks/Guilds/ether-swr/useGuildConfig';
+import { useGuildConfig } from 'hooks/Guilds/ether-swr/guild/useGuildConfig';
 
 const FilterContainer = styled(Box)`
   display: flex;
