+++ src/components/Guilds/ProposalStatus/index.tsx
@@ -6,7 +6,7 @@ import moment from 'moment';
 import { Box } from '../common/Layout';
 import { ProposalState } from '../../../types/types.guilds.d';
 import { useParams } from 'react-router';
-import { useProposal } from '../../../hooks/Guilds/ether-swr/useProposal';
+import { useProposal } from '../../../hooks/Guilds/ether-swr/guild/useProposal';
 
 const Status = styled.div`
   font-size: 0.8rem;
