+++ src/components/Guilds/StakeTokensModal/StakeTokens.tsx
@@ -9,8 +9,8 @@ import { useWeb3React } from '@web3-react/core';
 import { Heading } from '../common/Typography';
 import { Button } from '../common/Button';
 import dxIcon from '../../../assets/images/dxdao-icon.svg';
-import { useVotingPowerOf } from '../../../hooks/Guilds/ether-swr/useVotingPowerOf';
-import { useGuildConfig } from '../../../hooks/Guilds/ether-swr/useGuildConfig';
+import { useVotingPowerOf } from '../../../hooks/Guilds/ether-swr/guild/useVotingPowerOf';
+import { useGuildConfig } from '../../../hooks/Guilds/ether-swr/guild/useGuildConfig';
 import { useERC20Info } from '../../../hooks/Guilds/ether-swr/erc20/useERC20Info';
 import { useERC20Balance } from '../../../hooks/Guilds/ether-swr/erc20/useERC20Balance';
 import { formatUnits } from 'ethers/lib/utils';
